вынесла функцию в отдельный файл
