+++ helpers.php
@@ -142,5 +142,3 @@ function include_template($name, array $data = []) {
 
     return $result;
 }
-
-
